requirements.txt
fastapi==0.115.12
pymongo==4.13.2
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
uvicorn==0.34.3
pydantic[email]==2.11.7
bcrypt==4.0.1
email-validator==2.2.0
requests==2.31.0